python3Packages.shapely_1_8: remove
Shapely 1.8 is old unmaintained. We originally had to introduce it to nixpkgs in
order to package kikit, but recently kikit upstream dropped the
requirement to use an old version of shapely
(https://github.com/yaqwsx/KiKit/issues/574#issuecomment-1879721181).

Upstream relaxed the version constraint in
https://github.com/yaqwsx/KiKit/commit/3427a8d6ce31eea6fb3647e550d6b4ddc0e90842,
which first landed in kikit 1.5.0, which we're already packaging, so
we're free to drop shapely 1.8!